dary_heap

Struct DaryHeap

pub struct DaryHeap<T, const D: usize> { /* private fields */ }

A priority queue implemented with a d-ary heap.

This will be a max-heap.

It is a logic error for an item to be modified in such a way that the item’s ordering relative to any other item, as determined by the Ord trait, changes while it is in the heap. This is normally only possible through interior mutability, global state, I/O, or unsafe code. The behavior resulting from such a logic error is not specified, but will be encapsulated to the DaryHeap that observed the logic error and not result in undefined behavior. This could include panics, incorrect results, aborts, memory leaks, and non-termination.

As long as no elements change their relative order while being in the heap as described above, the API of DaryHeap guarantees that the heap invariant remains intact i.e. its methods all behave as documented. For example if a method is documented as iterating in sorted order, that’s guaranteed to work as long as elements in the heap have not changed order, even in the presence of closures getting unwinded out of, iterators getting leaked, and similar foolishness.

Usage

Rust type interference cannot infer the desired heap arity (value of d) automatically. Therefore, it is generally more ergonomic to use one of the type aliases instead of DaryHeap directly. See the crate-level documentation for more information.

Comparison to standard library

For a comparison with std::collections::BinaryHeap, see the crate-level documentation.

Examples

use dary_heap::BinaryHeap;

// Type inference lets us omit an explicit type signature (which
// would be `BinaryHeap<i32>` in this example).
let mut heap = BinaryHeap::new();

// We can use peek to look at the next item in the heap. In this case,
// there's no items in there yet so we get None.
assert_eq!(heap.peek(), None);

// Let's add some scores...
heap.push(1);
heap.push(5);
heap.push(2);

// Now peek shows the most important item in the heap.
assert_eq!(heap.peek(), Some(&5));

// We can check the length of a heap.
assert_eq!(heap.len(), 3);

// We can iterate over the items in the heap, although they are returned in
// a random order.
for x in &heap {
    println!("{x}");
}

// If we instead pop these scores, they should come back in order.
assert_eq!(heap.pop(), Some(5));
assert_eq!(heap.pop(), Some(2));
assert_eq!(heap.pop(), Some(1));
assert_eq!(heap.pop(), None);

// We can clear the heap of any remaining items.
heap.clear();

// The heap should now be empty.
assert!(heap.is_empty())

A DaryHeap with a known list of items can be initialized from an array:

use dary_heap::QuaternaryHeap;

let heap = QuaternaryHeap::from([1, 5, 2]);

Min-heap

Either core::cmp::Reverse or a custom Ord implementation can be used to make DaryHeap a min-heap. This makes heap.pop() return the smallest value instead of the greatest one.

use dary_heap::TernaryHeap;
use std::cmp::Reverse;

let mut heap = TernaryHeap::new();

// Wrap values in `Reverse`
heap.push(Reverse(1));
heap.push(Reverse(5));
heap.push(Reverse(2));

// If we pop these scores now, they should come back in the reverse order.
assert_eq!(heap.pop(), Some(Reverse(1)));
assert_eq!(heap.pop(), Some(Reverse(2)));
assert_eq!(heap.pop(), Some(Reverse(5)));
assert_eq!(heap.pop(), None);

Time complexity

push	pop	peek/peek_mut

O(1)~	O(log(n))	O(1)

The value for push is an expected cost; the method documentation gives a more detailed analysis.

Implementations

impl<T: Ord, const D: usize> DaryHeap<T, D>

pub fn new() -> DaryHeap<T, D>

Creates an empty DaryHeap as a max-heap.

Examples

Basic usage:

use dary_heap::QuaternaryHeap;
let mut heap = QuaternaryHeap::new();
heap.push(4);

pub fn with_capacity(capacity: usize) -> DaryHeap<T, D>

Creates an empty DaryHeap with at least the specific capacity.

The d-ary heap will be able to hold at least capacity elements without reallocating. This method is allowed to allocate for more elements than capacity. If capacity is 0, the d-ary heap will not allocate.

Examples

Basic usage:

use dary_heap::QuaternaryHeap;
let mut heap = QuaternaryHeap::with_capacity(10);
heap.push(4);

pub fn peek_mut(&mut self) -> Option<PeekMut<'_, T, D>>

Returns a mutable reference to the greatest item in the d-ary heap, or None if it is empty.

Note: If the PeekMut value is leaked, some heap elements might get leaked along with it, but the remaining elements will remain a valid heap.

Examples

Basic usage:

use dary_heap::TernaryHeap;
let mut heap = TernaryHeap::new();
assert!(heap.peek_mut().is_none());

heap.push(1);
heap.push(5);
heap.push(2);
{
    let mut val = heap.peek_mut().unwrap();
    *val = 0;
}
assert_eq!(heap.peek(), Some(&2));

Time complexity

If the item is modified then the worst case time complexity is O(log(n)), otherwise it’s O(1).

pub fn pop(&mut self) -> Option<T>

Removes the greatest item from the d-ary heap and returns it, or None if it is empty.

Examples

Basic usage:

use dary_heap::BinaryHeap;
let mut heap = BinaryHeap::from([1, 3]);

assert_eq!(heap.pop(), Some(3));
assert_eq!(heap.pop(), Some(1));
assert_eq!(heap.pop(), None);

Time complexity

The worst case cost of pop on a heap containing n elements is O(log(n)).

pub fn push(&mut self, item: T)

Pushes an item onto the d-ary heap.

Examples

Basic usage:

use dary_heap::QuaternaryHeap;
let mut heap = QuaternaryHeap::new();
heap.push(3);
heap.push(5);
heap.push(1);

assert_eq!(heap.len(), 3);
assert_eq!(heap.peek(), Some(&5));

Time complexity

The expected cost of push, averaged over every possible ordering of the elements being pushed, and over a sufficiently large number of pushes, is O(1). This is the most meaningful cost metric when pushing elements that are not already in any sorted pattern.

The time complexity degrades if elements are pushed in predominantly ascending order. In the worst case, elements are pushed in ascending sorted order and the amortized cost per push is O(log(n)) against a heap containing n elements.

The worst case cost of a single call to push is O(n). The worst case occurs when capacity is exhausted and needs a resize. The resize cost has been amortized in the previous figures.

pub fn into_sorted_vec(self) -> Vec<T>

Consumes the DaryHeap and returns a vector in sorted (ascending) order.

Examples

Basic usage:

use dary_heap::OctonaryHeap;

let mut heap = OctonaryHeap::from([1, 2, 4, 5, 7]);
heap.push(6);
heap.push(3);

let vec = heap.into_sorted_vec();
assert_eq!(vec, [1, 2, 3, 4, 5, 6, 7]);

pub fn append(&mut self, other: &mut Self)

Moves all the elements of other into self, leaving other empty.

Examples

Basic usage:

use dary_heap::OctonaryHeap;

let mut a = OctonaryHeap::from([-10, 1, 2, 3, 3]);
let mut b = OctonaryHeap::from([-20, 5, 43]);

a.append(&mut b);

assert_eq!(a.into_sorted_vec(), [-20, -10, 1, 2, 3, 3, 5, 43]);
assert!(b.is_empty());

pub fn retain<F>(&mut self, f: F)
where F: FnMut(&T) -> bool,

Retains only the elements specified by the predicate.

In other words, remove all elements e for which f(&e) returns false. The elements are visited in unsorted (and unspecified) order.

Examples

Basic usage:

use dary_heap::OctonaryHeap;

let mut heap = OctonaryHeap::from([-10, -5, 1, 2, 4, 13]);

heap.retain(|x| x % 2 == 0); // only keep even numbers

assert_eq!(heap.into_sorted_vec(), [-10, 2, 4])

impl<T, const D: usize> DaryHeap<T, D>

pub fn iter(&self) -> Iter<'_, T>

Returns an iterator visiting all values in the underlying vector, in arbitrary order.

Examples

Basic usage:

use dary_heap::TernaryHeap;
let heap = TernaryHeap::from([1, 2, 3, 4]);

// Print 1, 2, 3, 4 in arbitrary order
for x in heap.iter() {
    println!("{x}");
}

pub fn peek(&self) -> Option<&T>

Returns the greatest item in the d-ary heap, or None if it is empty.

Examples

Basic usage:

use dary_heap::BinaryHeap;
let mut heap = BinaryHeap::new();
assert_eq!(heap.peek(), None);

heap.push(1);
heap.push(5);
heap.push(2);
assert_eq!(heap.peek(), Some(&5));

Time complexity

Cost is O(1) in the worst case.

pub fn capacity(&self) -> usize

Returns the number of elements the d-ary heap can hold without reallocating.

Examples

Basic usage:

use dary_heap::OctonaryHeap;
let mut heap = OctonaryHeap::with_capacity(100);
assert!(heap.capacity() >= 100);
heap.push(4);

pub fn reserve_exact(&mut self, additional: usize)

Reserves the minimum capacity for at least additional elements more than the current length. Unlike reserve, this will not deliberately over-allocate to speculatively avoid frequent allocations. After calling reserve_exact, capacity will be greater than or equal to self.len() + additional. Does nothing if the capacity is already sufficient.

Panics

Panics if the new capacity overflows usize.

Examples

Basic usage:

use dary_heap::OctonaryHeap;
let mut heap = OctonaryHeap::new();
heap.reserve_exact(100);
assert!(heap.capacity() >= 100);
heap.push(4);

pub fn reserve(&mut self, additional: usize)

Reserves capacity for at least additional elements more than the current length. The allocator may reserve more space to speculatively avoid frequent allocations. After calling reserve, capacity will be greater than or equal to self.len() + additional. Does nothing if capacity is already sufficient.

Panics

Panics if the new capacity overflows usize.

Examples

Basic usage:

use dary_heap::BinaryHeap;
let mut heap = BinaryHeap::new();
heap.reserve(100);
assert!(heap.capacity() >= 100);
heap.push(4);

pub fn shrink_to_fit(&mut self)

Discards as much additional capacity as possible.

Examples

Basic usage:

use dary_heap::TernaryHeap;
let mut heap: TernaryHeap<i32> = TernaryHeap::with_capacity(100);

assert!(heap.capacity() >= 100);
heap.shrink_to_fit();
assert!(heap.capacity() == 0);

pub fn as_slice(&self) -> &[T]

Returns a slice of all values in the underlying vector, in arbitrary order.

Examples

Basic usage:

use dary_heap::OctonaryHeap;
use std::io::{self, Write};

let heap = OctonaryHeap::from([1, 2, 3, 4, 5, 6, 7]);

io::sink().write(heap.as_slice()).unwrap();

pub fn into_vec(self) -> Vec<T>

Consumes the DaryHeap and returns the underlying vector in arbitrary order.

Examples

Basic usage:

use dary_heap::QuaternaryHeap;
let heap = QuaternaryHeap::from([1, 2, 3, 4, 5, 6, 7]);
let vec = heap.into_vec();

// Will print in some order
for x in vec {
    println!("{x}");
}

pub fn len(&self) -> usize

Returns the length of the d-ary heap.

Examples

Basic usage:

use dary_heap::BinaryHeap;
let heap = BinaryHeap::from([1, 3]);

assert_eq!(heap.len(), 2);

pub fn is_empty(&self) -> bool

Checks if the d-ary heap is empty.

Examples

Basic usage:

use dary_heap::BinaryHeap;
let mut heap = BinaryHeap::new();

assert!(heap.is_empty());

heap.push(3);
heap.push(5);
heap.push(1);

assert!(!heap.is_empty());

pub fn drain(&mut self) -> Drain<'_, T>

Clears the d-ary heap, returning an iterator over the removed elements in arbitrary order. If the iterator is dropped before being fully consumed, it drops the remaining elements in arbitrary order.

The returned iterator keeps a mutable borrow on the heap to optimize its implementation.

Examples

Basic usage:

use dary_heap::QuaternaryHeap;
let mut heap = QuaternaryHeap::from([1, 3]);

assert!(!heap.is_empty());

for x in heap.drain() {
    println!("{x}");
}

assert!(heap.is_empty());

pub fn clear(&mut self)

Drops all items from the d-ary heap.

Examples

Basic usage:

use dary_heap::TernaryHeap;
let mut heap = TernaryHeap::from([1, 3]);

assert!(!heap.is_empty());

heap.clear();

assert!(heap.is_empty());